PostGresSQL/requirements.txt
python-dotenv
sqlalchemy
typer[all]
psycopg2-binary